// lib/types/products.ts
// lib/types/product.ts

/*
export type Product = {
  id?: number
  name: string
  src: string
  price: number | null
  images: { url: string; public_id: string | null }[]
  description: string
  in_stock: boolean
  category_id: string | null
  attributes: {
    sizes: string[]
    tags: string[]
    brand: string
    ratio: string
  }
}

export type ProductSummary = {
  id: number;
  name: string;
  images: { url: string; public_id: string | null }[];
};
*/